move big loop to structure
